BREAKING don't log into private repo if it's not necessary
!! THIS WILL BREAK EXISTING REPOS !!

Before this commit, the repos running the CI action had no choice but
always log into our private NPM repo using NPM_USER and NPM_PASS secrets.
If you didn't have those secrets, the CI would fail.
The error that caused the failure was that "NPM_PASS" is not defined
as a secret. Of course it's not, because there's no reason to log in
to any NPM repository in a public package, since npmjs is public.

After this commit, the CI action doesn't log in into verdaccio by default.
For repos like magento2 which are public, and don't use private
pacakges, this doesn't make sense.

Especially now that we use NPM_RELEASE_TOKEN secret for
releases (which allows you to publish without having to provide
NPM_USER and NPM_PASS), passing NPM_USER and NPM_PASS for a public
repo makes even less sense.

---

The breaking change comes from the fact that in private GH repos where
the CI workflow is reused without explicitly passing the param "enterprise:
true" (it's not passed because this argument didn't exist before this commit),
will now default to "enterprise: false".

This will break our private GitHub repos that need to `npm install`
private vsf packages to work. It will break because the CI NPM will
not be logged in.