BREAKING don't log into private repo if it's not necessary
!! THIS WILL BREAK EXISTING REPOS !!
since inputs.enterprise is false by default, npm won't log in to our
verdaccio and repos that are private and use our private verdaccio won't be able to download private enterprise
packages. The fix for this is adding enterprise: true to the workflow
in the affected repos.

Take for example the magento2 repository - it contains public packages that
rely on other public NPM packages. Logging in is not necessary.

This commit was made because trying to implement unified CI in
magento2 failed on the continuous-integration.yml run.

The error that caused the failure was that "NPM_PASS" is not defined
as a secret. Of course it's not, because there's no reason to log in
to any repository, since npmjs is public.