reduce max job run time in minutes from 360 to 120
Sometimes tests in integrations break and run for 6 hours. I'm worried
we'll cause the entire organization to run out of gh actions agent
time - yes it's finite and we pay for it.

Namely, the tests behave like this in sapcc, if you write the test
wrong the test process never returns an exit code and just hangs
until the timeout.

This commit is to reduce the damge done in case that happens.
120m is still a lot, but I'm being careful, knowing that other
teams (e.g. Unified Storefront) use those workflows.