Don't run husky precommit when changesets makes a release PR
Changesets makes a commit when it converts all the changeset files in
the <repo root>/.changesets folder into CHANGELOG.md

See:
https://github.com/changesets/changesets/blob/main/packages/release-utils/src/run.ts#L147

This is a problem if the repo that I'm running continuous-delivery.yml
in has Husky (git hooks, e.g. precommit) configured.

Husky will see that changesets is making an automated commit and it'll
just run the git precommit hook, which we don't want.

In this particular case that I'm fixing, the problem manifested itself
in the sap-commerce-cloud repository.

Wojtek was running a `ggshield`command. This is the GitGuardian executable that we're
already running as part of continuous-integration.yml

However, we don't install the `ggshield` executable into the PATH in
continuous-delivery, so the precommit hook failed. So the only
sensible thing to do was to disable the precommit hook *only in CD*.